third_party/WebKit/LayoutTests/fast/media/media-query-list-listener-expected.txt

Media query listeners should fire on changes to matches status.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaList.matches is true
PASS mediaList is mediaListArgument
PASS mediaList.matches is false
PASS mediaList is mediaListArgument
PASS mediaList.matches is true
PASS mediaList.matches is false
PASS successfullyParsed is true

TEST COMPLETE
